pyproject.toml
[tool.poetry]
name = "eventos2"
version = "0.1.0"
description = ""
authors = [
	"yurihs <yurisalvador@hotmail.com>"
]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
Django = "^3.0.2"
django-environ = "^0.4.5"
djangorestframework = "^3.10"
psycopg2-binary = "^2.8"
packaging = "^20.0"
Pillow = "^7.0.0"
gunicorn = "^20.0.4"
huey = "^2.1.2"
redis = "^3.3.11"
sentry-sdk = "^0.14.3"
rules = "^2.1"
pytest-django = "^3.7.0"
djangorestframework-simplejwt = "^4.4.0"
django-cors-headers = "^3.2.1"
python-magic = "^0.4.18"
drf-spectacular = "^0.9.11"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
black = "^19.10b0"
flake8 = "^3.7"
pre-commit = "^2.2.0"
toml = "^0.10.0"
flake8-bugbear = "^20.1.2"
isort = "^4.3"
bandit = "^1.6"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
